+++ src/components/ComplaintLandingPage/AskAuthority.astro
@@ -4,12 +4,14 @@ import Base from '../../layouts/base.astro';
 import { absUrl } from '../../lib/util';
 import TextRaw from '../TextRaw.astro';
 import { dpas } from '../../lib/dpas';
+import RadioForm from '../RadioForm.astro';
 
 interface Props {
     token: string;
+    complaintAuthority: string | null;
 }
 
-const { token } = Astro.props;
+const { token, complaintAuthority } = Astro.props;
 ---
 
 <Base title={t('complaint-landing-askAuthority', 'heading')}>
@@ -17,18 +19,13 @@ const { token } = Astro.props;
         <TextRaw scope="complaint-landing-askAuthority" id="question" />
     </p>
 
-    <div class="col66 col100-mobile">
-        <div class="radio-group radio-group-vertical radio-group-padded">
-            {
-                Object.values(dpas).map((d) => (
-                    <form class="radio-wrapper" method="POST" action={absUrl(`/p/${token}/complain/askAuthority`)}>
-                        <input type="hidden" name="answer" value={d.slug} />
-                        <button type="submit" class="radio-label">
-                            {d.name}
-                        </button>
-                    </form>
-                ))
-            }
-        </div>
-    </div>
+    <RadioForm
+        actionUrl={absUrl(`/p/${token}/complain/askAuthority/answer`)}
+        options={Object.keys(dpas).reduce((agg: Record<string, string>, key: string) => {
+            agg[key] = dpas[key]!.name;
+            return agg;
+        }, {})}
+        value={complaintAuthority}
+        idPrefix="askAuthority"
+    />
 </Base>
