+++ src/components/ComplaintLandingPage/AskDeviceHasRegisteredSimCard.astro
@@ -3,12 +3,14 @@ import { t } from '../../i18n/server';
 import Base from '../../layouts/base.astro';
 import { absUrl } from '../../lib/util';
 import TextRaw from '../TextRaw.astro';
+import RadioForm from '../RadioForm.astro';
 
 interface Props {
     token: string;
+    deviceHasRegisteredSimCard: boolean | null;
 }
 
-const { token } = Astro.props;
+const { token, deviceHasRegisteredSimCard } = Astro.props;
 ---
 
 <Base title={t('complaint-landing-askDeviceHasRegisteredSimCard', 'heading')}>
@@ -16,21 +18,16 @@ const { token } = Astro.props;
         <TextRaw scope="complaint-landing-askDeviceHasRegisteredSimCard" id="explanation" />
     </p>
 
-    <div class="col66 col100-mobile">
-        <div class="radio-group radio-group-vertical radio-group-padded">
-            {
-                (['yes', 'no'] as const).map((p) => (
-                    <form
-                        class="radio-wrapper"
-                        method="POST"
-                        action={absUrl(`/p/${token}/complain/askDeviceHasRegisteredSimCard`)}>
-                        <input type="hidden" name="answer" value={p} />
-                        <button type="submit" class="radio-label">
-                            {t('complaint-landing-askDeviceHasRegisteredSimCard', p)}
-                        </button>
-                    </form>
-                ))
-            }
-        </div>
-    </div>
+    <RadioForm
+        actionUrl={absUrl(`/p/${token}/complain/askDeviceHasRegisteredSimCard/answer`)}
+        options={(['yes', 'no'] as const).reduce(
+            (agg, p) =>
+                Object.assign(agg, {
+                    [p]: t('common', p),
+                }),
+            {},
+        )}
+        value={deviceHasRegisteredSimCard === null ? null : deviceHasRegisteredSimCard ? 'yes' : 'no'}
+        idPrefix="askDeviceHasRegisteredSimCard"
+    />
 </Base>
