+++ src/components/ComplaintLandingPage/AskLoggedIntoAppStore.astro
@@ -3,13 +3,15 @@ import { t } from '../../i18n/server';
 import Base from '../../layouts/base.astro';
 import { absUrl } from '../../lib/util';
 import TextRaw from '../TextRaw.astro';
+import RadioForm from '../RadioForm.astro';
 
 interface Props {
     token: string;
     platform: 'ios' | 'android';
+    loggedIntoAppStore: boolean | null;
 }
 
-const { token, platform } = Astro.props;
+const { token, platform, loggedIntoAppStore } = Astro.props;
 
 const store = t('common', `store-${platform}`);
 ---
@@ -19,21 +21,16 @@ const store = t('common', `store-${platform}`);
         <TextRaw scope="complaint-landing-askLoggedIntoAppStore" id="explanation" fields={{ store }} />
     </p>
 
-    <div class="col66 col100-mobile">
-        <div class="radio-group radio-group-vertical radio-group-padded">
-            {
-                (['yes', 'no'] as const).map((p) => (
-                    <form
-                        class="radio-wrapper"
-                        method="POST"
-                        action={absUrl(`/p/${token}/complain/askLoggedIntoAppStore`)}>
-                        <input type="hidden" name="answer" value={p} />
-                        <button type="submit" class="radio-label">
-                            {t('complaint-landing-askLoggedIntoAppStore', p, { store })}
-                        </button>
-                    </form>
-                ))
-            }
-        </div>
-    </div>
+    <RadioForm
+        actionUrl={absUrl(`/p/${token}/complain/askLoggedIntoAppStore/answer`)}
+        options={(['yes', 'no'] as const).reduce(
+            (agg, p) =>
+                Object.assign(agg, {
+                    [p]: t('complaint-landing-askLoggedIntoAppStore', p, { store }),
+                }),
+            {},
+        )}
+        value={loggedIntoAppStore === null ? null : loggedIntoAppStore ? 'yes' : 'no'}
+        idPrefix="askLoggedIntoAppStore"
+    />
 </Base>
